Update version stringification
Implement a way to specify different string formats for the version.  This is to be able to support varying version formatting for different systems; for example Docker has certain requirements where the version cannot contain a slash and sugarcrm needs to have a regularly incrementing value in order to install packages properly.